pyproject.toml
[tool.poetry]
name = "summarizer-agent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
cerebras-cloud-sdk = "^1.10.0"
pypdf2 = "^3.0.1"
python-multipart = "^0.0.17"
openai = "^1.54.4"
python-docx = "^1.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
